Use separate timeouts for each search.
For the DNS plugin, a change to the zone list means we need to rebuild
the zone files, but I don't want to do this twice if changes to both
zones and hosts come in at the same time. (This always happens if we
start with an empty cache.) Fix this by giving them separate timeouts,
and allowing plugins to invoke the timeout machinery in the same way as
LDAP change results. This means that zone list changes can have no
timeout, because they won't do any external work until the host list
timeout expires.