New DNS schema.
Include any additional records from the dNSZone entries. Currently the
list of valid record types to get from LDAP must be specified in the
config; I'm not sure if this is a useful security feature or a hack
because I can't be bothered to do it better...

dNSZone entries are now reserved for plain DNS RRs, and don't imply any
extra semantics. Nameservers are stored in mzNameserver entries. Each
mzNameserver represents a set of nameservers serving the same zones; the
master is named in the DN.

As well as the zoneName atts, these entries also have mzNetwork atts.
These values are looked up as networks(5) names, and any A/AAA record
generated from the hosts(5) entries with an address not in those
networks is dropped. Basically, 'this is the list of networks visible
from machines which will be using this nameserver'.